% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{clean_model_data}
\alias{clean_model_data}
\title{Internal method to process data}
\usage{
clean_model_data(formula, data, weights, condition_call, cluster_variable_name)
}
\description{
Internal method to process data
}
