% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_coef.R
\name{coef.difference_in_means}
\alias{coef.difference_in_means}
\title{Extract estimate from \code{\link{difference_in_means}}}
\usage{
\method{coef}{difference_in_means}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{difference_in_means}}}

\item{...}{other arguments, unused}
}
\description{
Extract estimate from \code{\link{difference_in_means}}
}
