% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.horvitz_thompson}
\alias{coef.horvitz_thompson}
\title{Extract estimate from \code{\link{horvitz_thompson}}}
\usage{
\method{coef}{horvitz_thompson}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{horvitz_thompson}}}

\item{...}{other arguments, unused}
}
\description{
Extract estimate from \code{\link{horvitz_thompson}}
}
