% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_starprep.R
\name{commarobust}
\alias{commarobust}
\title{Build lm_robust object from lm fit}
\usage{
commarobust(model, se_type, clusters = NULL, alpha = 0.05, ci = TRUE)
}
\description{
Build lm_robust object from lm fit
}
