% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_confint.R
\name{confint.difference_in_means}
\alias{confint.difference_in_means}
\title{Confidence intervals for \code{\link{difference_in_means}}objects}
\usage{
\method{confint}{difference_in_means}(object, parm = NULL, level = NULL,
  ...)
}
\arguments{
\item{object}{an object of class \code{\link{difference_in_means}}}

\item{parm}{unused}

\item{level}{the significance level, defaults to 'alpha' specified in \code{\link{difference_in_means}}}

\item{...}{other arguments, unused}
}
\description{
Confidence intervals for \code{\link{difference_in_means}}objects
}
