% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_pr_matrix.R
\name{declaration_to_condition_pr_mat}
\alias{declaration_to_condition_pr_mat}
\title{Build condition probaability matrix for Horvitz-Thompson estimation from randomization declaration}
\usage{
declaration_to_condition_pr_mat(declaration)
}
\arguments{
\item{declaration}{An object of class 'ra_declaration' that contains the experimental design}
}
\description{
Build condition probaability matrix for Horvitz-Thompson estimation from randomization declaration
}
