% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_clean_model_data.R
\name{demean_fes}
\alias{demean_fes}
\title{Demean it}
\usage{
demean_fes(model_data)
}
\arguments{
\item{model_data}{foo}
}
\description{
Demean it
}
\examples{

model_data = structure(list(c(`1` = 7, `2` = 2, `3` = 5, `4` = 3, `5` = 0,
`6` = 3, `7` = 2, `8` = 4, `9` = 7, `10` = 2, `11` = 4, `12` = 2,
`13` = 2, `14` = 6, `15` = 0, `16` = 4, `17` = 5, `18` = 7, `19` = 2,
`20` = 4, `21` = 4, `22` = 4, `23` = 4, `24` = 5, `25` = 3, `26` = 4,
`27` = 4, `28` = 9, `29` = 4, `30` = 3, `31` = 3, `32` = 2, `33` = 5,
`34` = 3, `35` = 6, `36` = 1, `37` = 4, `38` = 2, `39` = 9, `40` = 3
), structure(c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1,
1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 0,
0), .Dim = c(40L, 2L), .Dimnames = list(c("1", "2", "3", "4",
"5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15",
"16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
"27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37",
"38", "39", "40"), c("(Intercept)", "D")), assign = 0:1), structure(c(`1` = "d",
`2` = "a", `3` = "b", `4` = "a", `5` = "d", `6` = "d", `7` = "b",
`8` = "b", `9` = "d", `10` = "c", `11` = "d", `12` = "c", `13` = "c",
`14` = "a", `15` = "a", `16` = "c", `17` = "d", `18` = "c", `19` = "b",
`20` = "b", `21` = "d", `22` = "a", `23` = "b", `24` = "d", `25` = "c",
`26` = "d", `27` = "d", `28` = "d", `29` = "d", `30` = "c", `31` = "c",
`32` = "b", `33` = "a", `34` = "b", `35` = "c", `36` = "a", `37` = "c",
`38` = "a", `39` = "b", `40` = "c"), .Dim = c(40L, 1L), .Dimnames = list(
    NULL, "G")), structure(0, intercept=1)), .Names = c("outcome", "design_matrix",
"fixed_effects", "terms"))
str(demean_fes(model_data))
}
