% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_glance.R
\name{estimatr_glancers}
\alias{estimatr_glancers}
\alias{glance.lm_robust}
\title{Glance at an estimatr object}
\usage{
\method{glance}{lm_robust}(x, ...)
}
\arguments{
\item{x}{An object returned by one of the estimators}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with columns:
\item{r.squared}{The \eqn{R^2},
\deqn{R^2 = 1 - Sum(e[i]^2) / Sum((y[i] - y^*)^2),} where \eqn{y^*}
is the mean of \eqn{y[i]} if there is an intercept and zero otherwise,
and \eqn{e[i]} is the ith residual.}
\item{adj.r.squared}{The \eqn{R^2} but penalized for having more parameters, \code{rank}}
\item{statistic}{a vector with the value of the F-statistic with the numerator and denominator degrees of freedom}
\item{p.value}{p-value from the F test}
\item{df.residual}{residual degrees of freedom}
\item{N}{the number of observations used}
}
\description{
Glance at an estimatr object
}
\seealso{
\code{\link[generics:glance]{generics::glance()}}, \code{\link[estimatr:lm_robust]{estimatr::lm_robust()}}, \code{\link[estimatr:lm_lin]{estimatr::lm_lin()}}
}
\concept{estimatr glancers}
