% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_tidy.R
\name{estimatrZero_tidiers}
\alias{estimatrZero_tidiers}
\alias{tidy.lm_robust}
\alias{tidy.iv_robust}
\alias{tidy.difference_in_means}
\alias{tidy.horvitz_thompson}
\alias{tidy.lh_robust}
\alias{tidy.lh}
\title{Tidy an estimatrZero object}
\usage{
\method{tidy}{lm_robust}(x, conf.int = TRUE, conf.level = NULL, ...)

\method{tidy}{iv_robust}(x, conf.int = TRUE, conf.level = NULL, ...)

\method{tidy}{difference_in_means}(x, conf.int = TRUE, conf.level = NULL, ...)

\method{tidy}{horvitz_thompson}(x, conf.int = TRUE, conf.level = NULL, ...)

\method{tidy}{lh_robust}(x, conf.int = TRUE, conf.level = NULL, ...)

\method{tidy}{lh}(x, conf.int = TRUE, conf.level = NULL, ...)
}
\arguments{
\item{x}{An object returned by one of the estimators}

\item{conf.int}{Logical, whether to include confidence intervals.}

\item{conf.level}{The confidence level for intervals.}

\item{...}{extra arguments (not used)}
}
\description{
Tidy an estimatrZero object
}
