% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_extract.R
\name{extract}
\alias{extract}
\title{Extract model data for \pkg{texreg}}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{An object to be converged to a \pkg{texreg} object}

\item{...}{extra arguments}
}
\value{
a \pkg{texreg} objet
}
\description{
Extract model data for \pkg{texreg}
}
