% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_pr_matrix_complete}
\alias{gen_pr_matrix_complete}
\title{Gets a probability matrix for complete randomization from a vector of treatment probabilities}
\usage{
gen_pr_matrix_complete(prs)
}
\description{
Gets a probability matrix for complete randomization from a vector of treatment probabilities
}
