% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_robust.R
\name{lm_robust_se}
\alias{lm_robust_se}
\title{Ordinary Least Squares with Robust Standard Errors}
\usage{
lm_robust_se(formula, data, weights = NULL, alpha = 0.05, se_type = "HC2",
  coefficient_name = "Z")
}
\arguments{
\item{formula}{an object of class formula, as in \code{\link{lm}}.}

\item{data}{A data.frame.}

\item{weights}{the bare (unquoted) names of the weights variable in the supplied data.}

\item{alpha}{The significance level, 0.05 by default.}

\item{se_type}{The sort of standard error sought -- "HCO", "HC1", "HC2", "HC3", or "classical". "HC2" by default}

\item{coefficient_name}{a character or character vector that indicates which coefficients should be reported. Defaults to "Z".}
}
\description{
Ordinary Least Squares with Robust Standard Errors
}
