% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_na_omit_detailed.R
\name{na.omit_detailed.data.frame}
\alias{na.omit_detailed.data.frame}
\title{Extra logging on na.omit handler}
\usage{
na.omit_detailed.data.frame(object)
}
\arguments{
\item{object}{a data.frame}

\item{...}{unused}
}
\value{
a normal \code{omit} object, with the extra attribute \code{why_omit},
which contains the leftmost column containing an NA for each row that was dropped, by
column name, if any were dropped.
}
\description{
Extra logging on na.omit handler
}
\seealso{
\code{\link{na.omit}}
}
