% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nobs.R
\name{nobs.lm_robust}
\alias{nobs.lm_robust}
\title{Extract number of observations from \code{\link{lm_robust}} object}
\usage{
\method{nobs}{lm_robust}(object, ...)
}
\arguments{
\item{object}{an object of class 'lm_robust'}

\item{...}{other arguments, unused}
}
\description{
Extract number of observations from \code{\link{lm_robust}} object
}
