% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvitz_thompson_condition_pr_matrix.R
\name{permutations_to_condition_pr_mat}
\alias{permutations_to_condition_pr_mat}
\title{Build condition probability matrix for Horvitz-Thompson estimation from treatment permutations}
\usage{
permutations_to_condition_pr_mat(permutations)
}
\arguments{
\item{permutations}{A matrix where the rows are units and the columns are different treatment permutations; treated units must be represented with a 1 and control units with a 0}
}
\description{
Build condition probability matrix for Horvitz-Thompson estimation from treatment permutations
}
