% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.difference_in_means}
\alias{print.difference_in_means}
\title{Printing \code{\link{difference_in_means}} objects}
\usage{
\method{print}{difference_in_means}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{difference_in_means}}}

\item{...}{arguments passed to \code{\link{tidy.difference_in_means}}, unused}
}
\description{
Printing \code{\link{difference_in_means}} objects
}
