% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.horvitz_thompson}
\alias{print.horvitz_thompson}
\title{Printing \link{\code{horvitz_thompson}} objects}
\usage{
\method{print}{horvitz_thompson}(obj, ...)
}
\arguments{
\item{obj}{an object of class \link{\code{horvitz_thompson}}}

\item{...}{arguments passed to \link{\code{tidy.horvitz_thompson}}, unused}
}
\description{
Printing \link{\code{horvitz_thompson}} objects
}
