% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.lm_robust}
\alias{print.lm_robust}
\title{Printing \code{\link{lm_robust}} objects}
\usage{
\method{print}{lm_robust}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{lm_robust}}}

\item{...}{arguments passed to \code{\link{tidy.lm_robust}}, unused}
}
\description{
Printing \code{\link{lm_robust}} objects
}
