% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.summary.lm_robust}
\alias{print.summary.lm_robust}
\title{Printing \code{\link{summary.lm_robust}} objects}
\usage{
\method{print}{summary.lm_robust}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{\link{lm_robust}}}

\item{...}{arguments passed to \code{\link{tidy.lm_robust}}, unused}
}
\description{
Printing \code{\link{summary.lm_robust}} objects
}
