% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_declaration.R
\name{read_declaration}
\alias{read_declaration}
\title{Learns experimental design from randomization declaration}
\usage{
read_declaration(declaration, estimator)
}
\arguments{
\item{declaration}{An object of class 'ra_declaration' that contains the experimental design}

\item{estimator}{The estimator that needs to learn the experimental design}
}
\description{
Learns experimental design from randomization declaration
}
