% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_glance.R, R/S3_tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glance}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:glance]{glance()}}, \code{\link[generics:tidy]{tidy()}}}
}}

