% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_starprep.R
\name{starprep}
\alias{starprep}
\title{Prepare model fits for stargazer}
\usage{
starprep(..., stat = "se")
}
\arguments{
\item{...}{a list of lm_robust objects}

\item{stat}{either "se" (the default) or "p"}
}
\value{
a list of vectors of extracted statistics for stargazers
}
\description{
Prepare model fits for stargazer
}
\examples{

}
