% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.difference_in_means}
\alias{summary.difference_in_means}
\title{Summarizing \code{\link{difference_in_means}} objects}
\usage{
\method{summary}{difference_in_means}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{difference_in_means}}}

\item{...}{arguments passed to \code{\link{tidy.difference_in_means}}, unused}
}
\description{
Summarizing \code{\link{difference_in_means}} objects
}
