% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.horvitz_thompson}
\alias{summary.horvitz_thompson}
\title{Summarizing 'horvitz_thompson' objects}
\usage{
\method{summary}{horvitz_thompson}(obj, ...)
}
\arguments{
\item{obj}{an object of class 'horvitz_thompson'}

\item{...}{arguments passed to \link{\code{tidy.horvitz_thompson}}, unused}
}
\description{
Summarizing 'horvitz_thompson' objects
}
