% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lm_robust}
\alias{summary.lm_robust}
\title{Summarizing 'lm_robust' objects}
\usage{
\method{summary}{lm_robust}(obj, ...)
}
\arguments{
\item{obj}{an object of class 'lm_robust'}

\item{...}{arguments passed to \link{\code{tidy.lm_robust}}, unused}
}
\description{
Summarizing 'lm_robust' objects
}
