% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.default}
\alias{tidy.default}
\title{Default tidying method}
\usage{
\method{tidy}{default}(obj, ...)
}
\arguments{
\item{obj}{an object to be tidied}

\item{...}{extra arguments (not used)}
}
\description{
In order to be consistent with the broom package, the default tidy
method just tries to cast the object as a data frame
}
