% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.horvitz_thompson}
\alias{tidy.horvitz_thompson}
\title{Tidying horvitz_thompson output to a data.frame}
\usage{
\method{tidy}{horvitz_thompson}(obj, ...)
}
\arguments{
\item{obj}{a horvitz_thompson object to be tidied}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with with coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom
}
\description{
Tidying horvitz_thompson output to a data.frame
}
