% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.lm_robust}
\alias{tidy.lm_robust}
\title{Tidying lm_robust output to a data.frame}
\usage{
\method{tidy}{lm_robust}(object, ...)
}
\arguments{
\item{object}{a lm_robust object to be tidied}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with with coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom for the covariates
specified by which_covs
}
\description{
Tidying lm_robust output to a data.frame
}
