% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_tidy.R
\name{tidy.NULL}
\alias{tidy.NULL}
\title{tidy on a NULL input}
\usage{
\method{tidy}{NULL}(object, ...)
}
\arguments{
\item{object}{A value NULL}

\item{...}{extra arguments (not used)}
}
\value{
An empty data.frame
}
\description{
tidy on a NULL input returns an empty data frame, which means it can be
combined with other data frames (treated as "empty")
}
