% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy the result of a estimator into a data.frame}
\usage{
tidy(object, ...)
}
\arguments{
\item{object}{An object to be converted into a tidy data.frame}

\item{...}{extra arguments}
}
\value{
a data.frame
}
\description{
Tidy the result of a estimator into a data.frame
}
