% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{use_model}
\alias{use_model}
\title{Estimate and Summarize a Model}
\usage{
use_model(..., model = lm, data, alpha, coefficient_name = "Z")
}
\arguments{
\item{...}{arguments passed to model}

\item{model}{a modeling function like \code{\link{lm}} or \code{\link{glm}}.  Must have coef() and vcov() methods.}

\item{data}{A data.frame.}

\item{alpha}{The significance level, 0.05 by default. (not yet implemented)}

\item{coefficient_name}{a character or character vector that indicates which coefficients should be reported. Defaults to "Z".}
}
\description{
Estimate and Summarize a Model
}
