% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.difference_in_means}
\alias{vcov.difference_in_means}
\title{Return estimated variance covariance matrix of \link{\code{difference_in_means}}}
\usage{
\method{vcov}{difference_in_means}(object, ...)
}
\arguments{
\item{object}{An object of class \link{\code{difference_in_means}}}

\item{...}{extra arguments, unused}
}
\description{
Return estimated variance covariance matrix of \link{\code{difference_in_means}}
}
