% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.horvitz_thompson}
\alias{vcov.horvitz_thompson}
\title{Return estimated variance covariance matrix of horvitz_thompson}
\usage{
\method{vcov}{horvitz_thompson}(object, ...)
}
\arguments{
\item{object}{An object of class 'horvitz_thompson'}

\item{...}{extra arguments, unused}
}
\description{
Return estimated variance covariance matrix of horvitz_thompson
}
