% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.lm_lin}
\alias{vcov.lm_lin}
\title{Return variance covariance matrix of lm_lin object}
\usage{
\method{vcov}{lm_lin}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'lm_lin'}
}
\description{
Return variance covariance matrix of lm_lin object
}
