% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_predict.R
\name{augment.lm_robust}
\alias{augment.lm_robust}
\alias{augment.iv_robust}
\title{Augment a Model Object with Fitted Values and Residuals}
\usage{
\method{augment}{lm_robust}(x, data = NULL, newdata = NULL, ...)

\method{augment}{iv_robust}(x, data = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{an \code{lm_robust} or \code{iv_robust} object}

\item{data}{the data to augment, defaulting to the model frame}

\item{newdata}{optional new data to predict on instead}

\item{...}{ignored}
}
\value{
A \code{data.frame}.
}
\description{
Returns the model frame with \code{.fitted} and \code{.resid} columns appended, the
form downstream packages expect from \code{\link[broom:augment]{broom::augment()}}. Supplying
\code{newdata} returns that instead, with \code{.fitted} only.
}
