% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.horvitz_thompson}
\alias{coef.horvitz_thompson}
\title{Extract estimate from 'horvitz_thompson'}
\usage{
\method{coef}{horvitz_thompson}(obj, ...)
}
\arguments{
\item{obj}{an object of class 'horvitz_thompson'}

\item{...}{other arguments, unused}
}
\description{
Extract estimate from 'horvitz_thompson'
}
