% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_coef.R
\name{coef.lm_robust}
\alias{coef.lm_robust}
\title{Extract coefficients from \code{\link{lm_robust}} object}
\usage{
\method{coef}{lm_robust}(object, parm = NULL, ...)
}
\arguments{
\item{object}{an object of class 'lm_robust'}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. Returns all if missing.}

\item{...}{other arguments, unused}
}
\description{
Extract coefficients from \code{\link{lm_robust}} object
}
