% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.difference_in_means}
\alias{confint.difference_in_means}
\title{Confidence intervals for \link{\code{difference_in_means}}objects}
\usage{
\method{confint}{difference_in_means}(obj, level = NULL)
}
\arguments{
\item{obj}{an object of class 'difference_in_means'}

\item{level}{the significance level, defaults to 'alpha' specified in \link{\code{difference_in_means}}}
}
\description{
Confidence intervals for \link{\code{difference_in_means}}objects
}
