% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.horvitz_thompson}
\alias{confint.horvitz_thompson}
\title{Confidence intervals for \code{\link{horvitz_thompson}} objects}
\usage{
\method{confint}{horvitz_thompson}(object, parm = NULL, level = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link{horvitz_thompson}}}

\item{parm}{unused}

\item{level}{the significance level, defaults to 'alpha' specified in \code{\link{horvitz_thompson}}}

\item{...}{other arguments, unused}
}
\description{
Confidence intervals for \code{\link{horvitz_thompson}} objects
}
