% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_confint.R
\name{confint.lm_robust}
\alias{confint.lm_robust}
\title{Confidence intervals for  \code{\link{lm_robust}} objects}
\usage{
\method{confint}{lm_robust}(object, parm = NULL, level = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link{lm_robust}}}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. Returns all if missing.}

\item{level}{the significance level, defaults to 'alpha' specified in \code{\link{lm_robust}}}

\item{...}{other arguments, unused}
}
\description{
Confidence intervals for  \code{\link{lm_robust}} objects
}
