% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_in_means.R
\name{difference_in_means}
\alias{difference_in_means}
\title{Design-Based Difference-in-Means Estimator}
\usage{
difference_in_means(
  formula,
  data,
  blocks,
  clusters,
  weights,
  subset,
  se_type = c("default", "none"),
  condition1 = NULL,
  condition2 = NULL,
  ci = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{formula}{an object of class formula with one variable on the RHS}

\item{data}{A \code{data.frame}}

\item{blocks}{An optional bare (unquoted) name of the block variable}

\item{clusters}{An optional bare (unquoted) name of the cluster variable}

\item{weights}{the bare (unquoted) name of the weights variable}

\item{subset}{An optional bare (unquoted) expression specifying a subset}

\item{se_type}{\code{"default"} or \code{"none"}}

\item{condition1}{value in treatment for the control condition}

\item{condition2}{value in treatment for the treatment condition}

\item{ci}{logical. Whether to compute p-values and confidence intervals.}

\item{alpha}{The significance level, 0.05 by default.}
}
\value{
An object of class \code{"difference_in_means"}.
}
\description{
Design-Based Difference-in-Means Estimator
}
\details{
Selects the appropriate point estimate, standard errors, and degrees
of freedom for unit randomized, cluster randomized, block randomized,
block-cluster randomized, matched-pairs, and matched-pair cluster randomized
designs.

If weights are specified, estimation is handed to \code{\link[=lm_robust]{lm_robust()}} with HC2
standard errors.
}
\references{
Gerber, Alan S. and Donald P. Green. 2012. \emph{Field Experiments:
Design, Analysis, and Interpretation}. New York: W.W. Norton.

Imai, Kosuke, Gary King, and Clayton Nall. 2009. "The Essential Role of
Pair Matching in Cluster-Randomized Experiments." \emph{Statistical Science}
24(1): 29-53. \doi{10.1214/08-STS274}.
}
