% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_in_means.R
\name{difference_in_means}
\alias{difference_in_means}
\title{Built-in Estimators: Difference-in-means}
\usage{
difference_in_means(formula, condition1 = NULL, condition2 = NULL, data,
  weights = NULL, subset = NULL, alpha = 0.05)
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z}

\item{data}{A data.frame.}

\item{weights}{An optional vector of weights (not yet implemented).}

\item{subset}{An optional vector specifying a subset of observations to be used.}

\item{alpha}{The significance level, 0.05 by default.}
}
\description{
Built-in Estimators: Difference-in-means
}
