% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_in_means.R
\name{difference_in_means_blocked}
\alias{difference_in_means_blocked}
\title{Built-in Estimators: Blocked Difference-in-means}
\usage{
difference_in_means_blocked(formula, block_variable_name = NULL,
  condition1 = NULL, condition2 = NULL, data, weights = NULL,
  subset = NULL, alpha = 0.05)
}
\description{
Built-in Estimators: Blocked Difference-in-means
}
