% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatr-package.R
\docType{package}
\name{estimatr-package}
\alias{estimatr}
\alias{estimatr-package}
\title{estimatr: Fast Estimators for Design-Based Inference}
\description{
Fast procedures for a small set of commonly-used, design-appropriate
estimators with robust standard errors and confidence intervals. Provides
\code{lm_robust}, \code{lm_lin}, \code{iv_robust}, \code{lh_robust}, \code{difference_in_means}, and
\code{horvitz_thompson}, with feols-style fixed effects absorption and
design-aware Horvitz-Thompson variance. See \code{vignette("estimatr2.0")} for
what changes in version 2.0 and what does not.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://declaredesign.org/r/estimatr/}
  \item \url{https://github.com/DeclareDesign/estimatr}
  \item Report bugs at \url{https://github.com/DeclareDesign/estimatr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Coppock \email{acoppock@gmail.com}

Authors:
\itemize{
  \item Alexander Coppock \email{acoppock@gmail.com}
  \item Graeme Blair \email{graeme.blair@gmail.com}
  \item Jasper Cooper \email{jjc2247@columbia.edu}
  \item Luke Sonnet \email{luke.sonnet@gmail.com}
}

Other contributors:
\itemize{
  \item Macartan Humphreys \email{macartan@gmail.com} [contributor]
  \item Neal Fultz \email{nfultz@gmail.com} [contributor]
  \item Lily Medina \email{lilymiru@gmail.com} [contributor]
  \item Russell Lenth \email{russell-lenth@uiowa.edu} [contributor]
  \item Molly Offer-Westort \email{mollyow@uchicago.edu} [contributor]
}

}
