% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_tidy.R
\name{estimatr_tidiers}
\alias{estimatr_tidiers}
\alias{tidy.lm_robust}
\alias{tidy.iv_robust}
\alias{tidy.difference_in_means}
\alias{tidy.horvitz_thompson}
\alias{tidy.lh_robust}
\alias{tidy.lh}
\title{Tidy an estimatr object}
\usage{
\method{tidy}{lm_robust}(x, ...)

\method{tidy}{iv_robust}(x, ...)

\method{tidy}{difference_in_means}(x, ...)

\method{tidy}{horvitz_thompson}(x, ...)

\method{tidy}{lh_robust}(x, ...)

\method{tidy}{lh}(x, ...)
}
\arguments{
\item{x}{An object returned by one of the estimators}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with columns for coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom, and the
name of the outcome variable
}
\description{
Tidy an estimatr object
}
\seealso{
\code{\link[generics:tidy]{generics::tidy()}}, \code{\link[estimatr:lm_robust]{estimatr::lm_robust()}}, \code{\link[estimatr:iv_robust]{estimatr::iv_robust()}},  \code{\link[estimatr:difference_in_means]{estimatr::difference_in_means()}}, \code{\link[estimatr:horvitz_thompson]{estimatr::horvitz_thompson()}}
}
\concept{estimatr tidiers}
