% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatrZero.R
\docType{package}
\name{estimatrZero-package}
\alias{estimatrZero}
\alias{estimatrZero-package}
\title{estimatrZero: Fast Estimators for Design-Based Inference}
\description{
A lean rewrite of estimatr targeting the DeclareDesign workflow. Provides
\code{lm_robust}, \code{lm_lin}, \code{iv_robust}, \code{lh_robust}, \code{difference_in_means}, and
\code{horvitz_thompson}, with feols-style fixed effects absorption and
design-aware Horvitz-Thompson variance. See \code{vignette("estimatr2.0")} for
what changes relative to estimatr and what does not.
}
\author{
\strong{Maintainer}: Alex Coppock \email{acoppock@gmail.com}

Authors:
\itemize{
  \item Alex Coppock \email{acoppock@gmail.com}
}

}
