% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatrZero.R
\docType{package}
\name{estimatrZero-package}
\alias{estimatrZero}
\alias{estimatrZero-package}
\title{estimatrZero: Fast Estimators for Design-Based Inference}
\description{
A lean rewrite of estimatr. Provides \code{lm_robust}, \code{lm_lin}, \code{iv_robust},
\code{lh_robust}, and \code{difference_in_means} with HC2/CR2 standard errors.
Fixed effects and Horvitz-Thompson are intentionally omitted.
Weighted R-squared bug fixed.
}
\author{
\strong{Maintainer}: Alex Coppock \email{acoppock@gmail.com}

}
