% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_extract.R
\name{extract.lm_robust}
\alias{extract.lm_robust}
\title{Extract model data for texreg output}
\usage{
extract.lm_robust(model, include.ci = TRUE, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.fstatistic = FALSE,
  include.rmse = TRUE, ...)
}
\arguments{
\item{model}{an object of class \code{\link{lm_robust}}}
}
\description{
Prepares an \code{"lm_robust"} object for the \pkg{texreg}
package.
}
