% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_extract.R
\name{extract.robust_default}
\alias{extract.robust_default}
\alias{extract.lm_robust}
\alias{extract.iv_robust}
\title{Extract model data for \pkg{texreg} package}
\usage{
extract.robust_default(model, include.ci = TRUE, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.fstatistic = FALSE,
  include.rmse = TRUE, ...)

extract.lm_robust(model, include.ci = TRUE, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.fstatistic = FALSE,
  include.rmse = TRUE, ...)

extract.iv_robust(model, include.ci = TRUE, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.fstatistic = FALSE,
  include.rmse = TRUE, ...)
}
\arguments{
\item{model}{an object of class \code{\link{lm_robust}} or \code{"iv_robust"}}

\item{include.ci}{logical. Defaults to TRUE}

\item{include.rsquared}{logical. Defaults to TRUE}

\item{include.adjrs}{logical. Defaults to TRUE}

\item{include.nobs}{logical. Defaults to TRUE}

\item{include.fstatistic}{logical. Defaults to TRUE}

\item{include.rmse}{logical. Defaults to TRUE}

\item{...}{unused}
}
\description{
Prepares a \code{"lm_robust"} or \code{"iv_robust"} object for the \pkg{texreg}
package. This is largely a clone of the \code{extract.lm}
method.
}
