% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_declaration.R
\name{get_condition_pr_matrix}
\alias{get_condition_pr_matrix}
\title{Wrapper to just return condition_pr_matrix to speed up horvitz_thompson}
\usage{
get_condition_pr_matrix(declaration)
}
\arguments{
\item{declaration}{An object of class 'ra_declaration' that contains the experimental design}
}
\description{
Wrapper to just return condition_pr_matrix to speed up horvitz_thompson
}
