% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvitz_thompson.R
\name{horvitz_thompson}
\alias{horvitz_thompson}
\title{Horvitz-Thompson estimator of treatment effects}
\usage{
horvitz_thompson(formula, data, condition_pr_variable_name, block_variable_name,
  cluster_variable_name, condition_pr_matrix = NULL, declaration = NULL,
  se_type = c("youngs", "constant"), subset = NULL, alpha = 0.05,
  condition1 = NULL, condition2 = NULL, estimator = "ht")
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z}

\item{data}{A data.frame.}

\item{condition_pr_variable_name}{A bare (unquoted) name of the variable with the condition 2 (treatment) probabilities.}

\item{block_variable_name}{An optional bare (unquoted) name of the block variable. Use for blocked designs only.}

\item{cluster_variable_name}{An optional bare (unquoted) name of the variable that corresponds to the clusters in the data; used for cluster randomized designs. For blocked designs, clusters must be within blocks.}

\item{condition_pr_matrix}{A 2n * 2n matrix of marginal and joint probabilities of all units in condition1 and condition2}

\item{declaration}{An object of class "ra_declaration", from the randomizr package that is an alternative way of specifying the design. Cannot be used along with any of \code{condition_pr_variable_name}, \code{block_variable_name}, \code{cluster_variable_name}, or \code{condition_pr_matrix}.}

\item{se_type}{can be one of \code{c("youngs", "constant")} and correspond's to estimating the standard errors using Young's inequality (default, conservative), or the constant effects assumption.}

\item{subset}{An optional bare (unquoted) expression specifying a subset of observations to be used.}

\item{alpha}{The significance level, 0.05 by default.}

\item{condition1}{names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.}

\item{condition2}{names of the conditions to be compared. Effects are estimated with condition1 as control and condition2 as treatment. If unspecified, condition1 is the "first" condition and condition2 is the "second" according to r defaults.}
}
\description{
Horvitz-Thompson estimator of treatment effects
}
\details{
This function implements the Horvitz-Thompson estimator for treatment effects.
}
