% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_robust.R
\name{iv_robust}
\alias{iv_robust}
\title{Two-Stage Least Squares Instrumental Variables Regression}
\usage{
iv_robust(
  formula,
  data,
  weights,
  subset,
  clusters,
  se_type = NULL,
  ci = TRUE,
  alpha = 0.05,
  diagnostics = FALSE,
  return_vcov = TRUE,
  try_cholesky = FALSE
)
}
\arguments{
\item{formula}{an object of class formula with regressors and instruments,
e.g. \code{y ~ x1 + x2 | z1 + z2}.}

\item{data}{A \code{data.frame}}

\item{weights}{the bare (unquoted) name of the weights variable}

\item{subset}{An optional bare (unquoted) expression specifying a subset}

\item{clusters}{An optional bare (unquoted) name of the cluster variable}

\item{se_type}{The sort of standard error (see \code{\link[=lm_robust]{lm_robust()}})}

\item{ci}{logical. Whether to compute p-values and confidence intervals.}

\item{alpha}{The significance level, 0.05 by default.}

\item{diagnostics}{logical. Whether to compute IV diagnostic statistics.}

\item{return_vcov}{logical. Whether to return the vcov matrix.}

\item{try_cholesky}{logical. Whether to try Cholesky decomposition.}
}
\value{
An object of class \code{"iv_robust"}.
}
\description{
Two-Stage Least Squares Instrumental Variables Regression
}
