% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh_robust.R
\name{lh_robust}
\alias{lh_robust}
\title{Linear Hypothesis Test for OLS with Robust Standard Errors}
\usage{
lh_robust(..., data, linear_hypothesis)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[=lm_robust]{lm_robust()}}}

\item{data}{A \code{data.frame}}

\item{linear_hypothesis}{A character string or matrix specifying the
hypothesis, passed to \code{car::linearHypothesis}}
}
\value{
An object of class \code{"lh_robust"} with components \code{lm_robust}, \code{lh},
and \code{joint_hypothesis}.
}
\description{
Linear Hypothesis Test for OLS with Robust Standard Errors
}
