% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.R
\name{lm_fit}
\alias{lm_fit}
\title{Internal method that creates linear fits}
\usage{
lm_fit(y, design_matrix, weights, cluster, ci, se_type, alpha, coefficient_name)
}
\description{
Internal method that creates linear fits
}
