% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lin.R
\name{lm_lin}
\alias{lm_lin}
\title{Linear Regression with Lin (2013) Covariate Adjustment}
\usage{
lm_lin(
  formula,
  covariates,
  data,
  weights,
  subset,
  clusters,
  se_type = NULL,
  ci = TRUE,
  alpha = 0.05,
  return_vcov = TRUE,
  try_cholesky = FALSE
)
}
\arguments{
\item{formula}{an object of class formula with only the treatment on the RHS}

\item{covariates}{a right-sided formula with pre-treatment covariates}

\item{data}{A \code{data.frame}}

\item{weights}{the bare (unquoted) name of the weights variable}

\item{subset}{An optional bare (unquoted) expression specifying a subset}

\item{clusters}{An optional bare (unquoted) name of the cluster variable}

\item{se_type}{The sort of standard error (see \code{\link[=lm_robust]{lm_robust()}})}

\item{ci}{logical. Whether to compute p-values and confidence intervals.}

\item{alpha}{The significance level, 0.05 by default.}

\item{return_vcov}{logical. Whether to return the vcov matrix.}

\item{try_cholesky}{logical. Whether to try Cholesky decomposition.}
}
\value{
An object of class \code{"lm_robust"}.
}
\description{
Linear Regression with Lin (2013) Covariate Adjustment
}
\references{
Lin, Winston. 2013. "Agnostic Notes on Regression Adjustments to
Experimental Data: Reexamining Freedman's Critique." The Annals of Applied
Statistics 7(1): 295-318. \doi{10.1214/12-AOAS583}.
}
