% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lin.R
\name{lm_lin}
\alias{lm_lin}
\title{Linear regression with the Lin (2013) covariate adjustment}
\usage{
lm_lin(formula, data, covariates, weights = NULL, subset = NULL,
  cluster_variable_name = NULL, se_type = NULL, ci = TRUE, alpha = 0.05,
  coefficient_name = NULL)
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z. Should only have the outcome and the treatment.}

\item{data}{A data.frame.}

\item{covariates}{A one-sided formula with all of the covariates on the right hand side, such as ~ x1 + x2 + x3.}

\item{...}{All other arguments that are passed through to \code{\link{lm_robust}}}
}
\description{
Linear regression with the Lin (2013) covariate adjustment
}
