% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lin.R
\name{lm_lin}
\alias{lm_lin}
\title{A Wrapper for Lin 2013 Covariate Adjustment for OLS with Robust Standard Errors}
\usage{
lm_lin(formula, data, covariates, subset = NULL, ...)
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z. Should only have the outcome and the treatment.}

\item{data}{A data.frame.}

\item{covariates}{A one-sided formula with all of the covariates on the right hand side, such as ~ x1 + x2 + x3.}

\item{...}{All other arguments that are passed through to \code{\link{lm_robust}}}
}
\description{
A Wrapper for Lin 2013 Covariate Adjustment for OLS with Robust Standard Errors
}
