% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_lin.R
\name{lm_lin}
\alias{lm_lin}
\title{Linear regression with the Lin (2013) covariate adjustment}
\usage{
lm_lin(formula, data, covariates, weights, subset, cluster_variable_name,
  se_type = NULL, ci = TRUE, alpha = 0.05, coefficient_name = NULL,
  return_vcov = TRUE)
}
\arguments{
\item{formula}{An object of class "formula", such as Y ~ Z. Should only have the outcome and the treatment.}

\item{data}{A data.frame.}

\item{covariates}{A one-sided formula with all of the covariates on the right hand side, such as ~ x1 + x2 + x3.}

\item{weights}{the bare (unquoted) names of the weights variable in the supplied data.}

\item{subset}{An optional bare (unquoted) expression specifying a subset of observations to be used.}

\item{cluster_variable_name}{An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.}

\item{se_type}{The sort of standard error sought. Without clustering: "HC0", "HC1", "HC2" (default), "HC3", or "classical". With clustering: "BM" (default), "stata".}

\item{ci}{A boolean for whether to compute and return pvalues and confidence intervals, TRUE by default.}

\item{alpha}{The significance level, 0.05 by default.}

\item{coefficient_name}{a character or character vector that indicates which coefficients should be reported. If left unspecified, returns all coefficients.}

\item{return_vcov}{a boolean for whether to return the vcov matrix for later usage, TRUE by default.}
}
\description{
Linear regression with the Lin (2013) covariate adjustment
}
