% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_robust.R
\name{lm_robust}
\alias{lm_robust}
\title{Ordinary Least Squares with Robust Standard Errors}
\usage{
lm_robust(
  formula,
  data,
  weights,
  subset,
  clusters,
  se_type = NULL,
  ci = TRUE,
  alpha = 0.05,
  return_vcov = TRUE,
  try_cholesky = FALSE
)
}
\arguments{
\item{formula}{an object of class formula, as in \code{\link[=lm]{lm()}}}

\item{data}{A \code{data.frame}}

\item{weights}{the bare (unquoted) name of the weights variable}

\item{subset}{An optional bare (unquoted) expression specifying a subset}

\item{clusters}{An optional bare (unquoted) name of the cluster variable}

\item{se_type}{The sort of standard error. Without clusters: "HC0", "HC1",
"HC2" (default), "HC3", "classical", "stata", or "none". With clusters:
"CR0", "CR2" (default), "stata", or "none".}

\item{ci}{logical. Whether to compute p-values and confidence intervals.}

\item{alpha}{The significance level, 0.05 by default.}

\item{return_vcov}{logical. Whether to return the vcov matrix.}

\item{try_cholesky}{logical. Whether to try Cholesky decomposition.}
}
\value{
An object of class \code{"lm_robust"}.
}
\description{
Ordinary Least Squares with Robust Standard Errors
}
