% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatr_lm_robust.R
\name{lm_robust}
\alias{lm_robust}
\title{Ordinary Least Squares with Robust Standard Errors}
\usage{
lm_robust(formula, data, weights, subset, clusters, se_type = NULL,
  ci = TRUE, alpha = 0.05, coefficient_name = NULL, return_vcov = TRUE,
  try_cholesky = FALSE)
}
\arguments{
\item{formula}{an object of class formula, as in \code{\link{lm}}.}

\item{data}{A data.frame.}

\item{weights}{the bare (unquoted) names of the weights variable in the supplied data.}

\item{subset}{An optional bare (unquoted) expression specifying a subset of observations to be used.}

\item{clusters}{An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.}

\item{se_type}{The sort of standard error sought. Without clustering: "HC0", "HC1", "HC2" (default), "HC3", or "classical". With clustering: "CR2" (default), "stata".}

\item{ci}{A boolean for whether to compute and return pvalues and confidence intervals, TRUE by default.}

\item{alpha}{The significance level, 0.05 by default.}

\item{coefficient_name}{a character or character vector that indicates which coefficients should be reported. If left unspecified, returns all coefficients.}

\item{return_vcov}{a boolean for whether to return the vcov matrix for later usage, TRUE by default.}

\item{try_cholesky}{a boolean for whether to try using a cholesky decomposition to solve LS instead of a QR decomposition, FALSE by default. See 'details'.}
}
\description{
Ordinary Least Squares with Robust Standard Errors
}
