% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_robust.R
\name{lm_robust}
\alias{lm_robust}
\title{Ordinary Least Squares with Robust Standard Errors}
\usage{
lm_robust(formula, data, weights = NULL, subset = NULL,
  cluster_variable_name = NULL, se_type = NULL, ci = TRUE, alpha = 0.05,
  coefficient_name = NULL)
}
\arguments{
\item{formula}{an object of class formula, as in \code{\link{lm}}.}

\item{data}{A data.frame.}

\item{weights}{the bare (unquoted) names of the weights variable in the supplied data.}

\item{subset}{An optional bare (unquoted) expression specifying a subset of observations to be used.}

\item{cluster_variable_name}{An optional bare (unquoted) name of the variable that corresponds to the clusters in the data.}

\item{se_type}{The sort of standard error sought. Without clustering: "HC0", "HC1", "HC2" (default), "HC3", or "classical". With clustering: "BM" (default), "stata".}

\item{ci}{A boolean for whether to compute and return pvalues and confidence intervals, TRUE by default.}

\item{alpha}{The significance level, 0.05 by default.}

\item{coefficient_name}{a character or character vector that indicates which coefficients should be reported. If left unspecified, returns all coefficients.}
}
\description{
Ordinary Least Squares with Robust Standard Errors
}
