% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lm_robust_fit.R
\name{lm_robust_fit}
\alias{lm_robust_fit}
\title{Internal method that creates linear fits}
\usage{
lm_robust_fit(y, X, weights, cluster, ci, se_type, alpha, return_vcov,
  try_cholesky, has_int)
}
\arguments{
\item{y}{numeric outcome vector}

\item{X}{numeric design matrix}

\item{weights}{numeric weights vector}

\item{cluster}{numeric cluster vector}

\item{ci}{boolean that when T returns confidence intervals and p-values}

\item{se_type}{character denoting which kind of SEs to return}

\item{alpha}{numeric denoting the test size for confidence intervals}

\item{return_vcov}{a boolean for whether to return the vcov matrix for later usage}

\item{try_cholesky}{a boolean for whether to try using a cholesky decomposition to solve LS instead of a QR decomposition}

\item{has_int}{a boolean for whether the model has an intercept}
}
\description{
Internal method that creates linear fits
}
