% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_lm_robust_fit.R
\name{lm_robust_fit}
\alias{lm_robust_fit}
\title{Internal method that creates linear fits}
\usage{
lm_robust_fit(y, X, weights, cluster, ci, se_type, has_int, alpha = 0.05,
  return_vcov = TRUE, return_fit = FALSE, return_unweighted_fit = FALSE,
  try_cholesky = FALSE, X_first_stage = NULL)
}
\arguments{
\item{y}{numeric outcome vector}

\item{X}{numeric design matrix}

\item{weights}{numeric weights vector}

\item{cluster}{numeric cluster vector}

\item{ci}{boolean that when T returns confidence intervals and p-values}

\item{se_type}{character denoting which kind of SEs to return}

\item{has_int}{logical, whether the model has an intercept, used for \eqn{R^2}}

\item{alpha}{numeric denoting the test size for confidence intervals}

\item{return_vcov}{logical, whether to return the vcov matrix for later usage}

\item{return_fit}{logical, whether to return fitted values}

\item{return_unweighted_fit}{logical, whether to return unweighted fitted values in place of weighted fitted values if regression is weighted}

\item{try_cholesky}{logical, whether to try using a cholesky decomposition to solve LS instead of a QR decomposition}

\item{X_first_stage}{numeric matrix of the first stage design matrix, only use for second stage of 2SLS IV regression, otherwise leave as \code{NULL}}
}
\description{
Internal method that creates linear fits
}
