% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lm_robust}
\alias{predict.lm_robust}
\title{Model predictions for linear models}
\usage{
\method{predict}{lm_robust}(obj, newdata, se.fit = FALSE,
  interval = c("none", "confidence", ""))
}
\arguments{
\item{obj}{An object of class 'lm_robust'}

\item{level}{}
}
\description{
Model predictions for linear models
}
