% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lm_robust}
\alias{predict.lm_robust}
\title{Predict method for \code{\link{lm_robust}} object}
\usage{
\method{predict}{lm_robust}(object, newdata, se.fit = FALSE,
  interval = c("none", "confidence", "prediction"), level = 0.95,
  na.action = na.pass, ...)
}
\arguments{
\item{object}{an object of class 'lm_robust'}

\item{newdata}{the data for which a prediction is preferred}

\item{se.fit}{a boolean for whether standard errors are required, default = FALSE}

\item{na.action}{function determining what should be done with missing values in newdata. The default is to predict NA.}

\item{...}{other arguments, unused}
}
\description{
Predict method for \code{\link{lm_robust}} object
}
