% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.difference_in_means}
\alias{print.difference_in_means}
\title{Printing \link{\code{difference_in_means}} objects}
\usage{
\method{print}{difference_in_means}(obj, ...)
}
\arguments{
\item{obj}{an object of class \link{\code{difference_in_means}}}

\item{...}{arguments passed to \link{\code{tidy.difference_in_means}}, unused}
}
\description{
Printing \link{\code{difference_in_means}} objects
}
