% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.horvitz_thompson}
\alias{print.horvitz_thompson}
\title{Printing \code{\link{horvitz_thompson}} objects}
\usage{
\method{print}{horvitz_thompson}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{horvitz_thompson}}}

\item{...}{arguments passed to \code{\link{tidy.horvitz_thompson}}, unused}
}
\description{
Printing \code{\link{horvitz_thompson}} objects
}
