% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lm_robust}
\alias{print.lm_robust}
\title{Printing \link{\code{lm_robust}} objects}
\usage{
\method{print}{lm_robust}(obj, ...)
}
\arguments{
\item{obj}{an object of class \link{\code{lm_robust}}}

\item{...}{arguments passed to \link{\code{tidy.lm_robust}}, unused}
}
\description{
Printing \link{\code{lm_robust}} objects
}
