% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.difference_in_means}
\alias{summary.difference_in_means}
\title{Summarizing 'difference_in_means' objects}
\usage{
\method{summary}{difference_in_means}(obj, ...)
}
\arguments{
\item{obj}{an object of class 'difference_in_means'}

\item{...}{arguments passed to \link{\code{tidy.difference_in_means}}, unused}
}
\description{
Summarizing 'difference_in_means' objects
}
