% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.horvitz_thompson}
\alias{summary.horvitz_thompson}
\title{Summarizing \code{\link{horvitz_thompson}} objects}
\usage{
\method{summary}{horvitz_thompson}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{horvitz_thompson}}}

\item{...}{arguments passed to \code{\link{tidy.horvitz_thompson}}, unused}
}
\description{
Summarizing \code{\link{horvitz_thompson}} objects
}
