% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.lm_robust}
\alias{summary.lm_robust}
\title{Summarizing \code{\link{lm_robust}} objects}
\usage{
\method{summary}{lm_robust}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{lm_robust}}}

\item{...}{arguments passed to \code{\link{tidy.lm_robust}}, unused}
}
\description{
Summarizing \code{\link{lm_robust}} objects
}
