% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.difference_in_means}
\alias{tidy.difference_in_means}
\title{Tidying difference_in_means output to a data.frame}
\usage{
\method{tidy}{difference_in_means}(object, ...)
}
\arguments{
\item{object}{a horvitz_thompson object to be tidied}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with with coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom
}
\description{
Tidying difference_in_means output to a data.frame
}
