% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.difference_in_means}
\alias{vcov.difference_in_means}
\alias{vcov.difference_in_means}
\title{Return estimated variance covariance matrix of difference_in_means}
\usage{
\method{vcov}{difference_in_means}(obj)

\method{vcov}{difference_in_means}(obj)
}
\arguments{
\item{obj}{An object of class 'difference_in_means'}

\item{...}{extra arguments, unused}

\item{obj}{An object of class 'horvitz_thompson'}

\item{...}{extra arguments, unused}
}
\description{
Return estimated variance covariance matrix of difference_in_means

Return estimated variance covariance matrix of horvitz_thompson
}
