% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.lm_robust}
\alias{vcov.lm_robust}
\title{Return estimated variance covariance matrix of \link{\code{lm_robust}}}
\usage{
\method{vcov}{lm_robust}(object, ...)
}
\arguments{
\item{object}{An object of class \link{\code{lm_robust}}}

\item{...}{extra arguments, unused}
}
\description{
Return estimated variance covariance matrix of \link{\code{lm_robust}}
}
