% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.lm_robust}
\alias{vcov.lm_robust}
\title{Return variance covariance matrix of lm_robust object}
\usage{
\method{vcov}{lm_robust}(obj, ...)
}
\arguments{
\item{obj}{An object of class 'lm_robust'}
}
\description{
Return variance covariance matrix of lm_robust object
}
